backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.2",
    "datasets>=4.5.0",
    "einops>=0.8.2",
    "fastapi>=0.129.0",
    "google-genai>=1.64.0",
    "groq>=1.0.0",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "jupyterlab>=4.5.4",
    "motor>=3.7.1",
    "networkx>=3.6.1",
    "notebook>=7.5.3",
    "numpy>=2.4.2",
    "pandas>=3.0.1",
    "peft>=0.18.1",
    "protobuf>=6.33.5",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymongo>=4.16.0",
    "pymupdf>=1.27.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "safetensors>=0.7.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "sentencepiece>=0.2.1",
    "spacy>=3.8.11",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.2.0",
    "trl>=0.28.0",
    "uvicorn[standard]>=0.41.0",
]
